Add detailed comments to existing Java classes, enhance readability
This commit enhances the readability of the Minecraft server plugin. It does so by adding detailed comments to describe the functionalities of the classes, methods, and the parameters used in SurvivalWorldCommand, SetHubCommand, SetSurvivalWorldCommand, and other classes. The reason for this change is to make the code easier to understand and maintain. Furthermore, it will help new developers to get up to speed quicker with the project. New package-info files have also been created for providing package-level documentation.