Added descriptive comments to command classes
Comments have been added to the command classes (SurvivalWorldCommand, SetCommand, TeleportCommand, HubCommand) to enhance readability and understanding of the code. This includes high-level explanations of each class, detail on the methods within them and their functionality, as well as clarifications on the parameters and return values where necessary. Now, newcomers to the code or developers revisiting it should have a clearer understanding of the purpose and workings of these classes.