Updated build version to 1.6.0
The build version of the LilithServerEssential has been updated from 1.5.2 to 1.6.0, due to the addition of a new feature that checks the server's version on startup. This feature will disable the plugin and log an error if the server is running an unsupported version or if the version cannot be determined, thus enhancing the plugin's stability and compatibility.