Remove server version check from onEnable()
Deleted the server's version check code from the onEnable() method in the LilithServerEssential.java file. This code led to issues with the plugin's compatibility with varying server versions, making it unreliable and redundant. To improve adaptability across various server versions, it was decided to remove it. Additionally, the build version has been updated to 1.6.1 in the build.gradle file.