Rename classes and namespace consistent with language conventions
The primary change was to comply with the naming conventions of Java and improve readability. The names of the various java classes and the package names have been renamed by using underscores(_). As java does not conventionally use underscores in class or package names, this change has been reverted to improve readability and maintainability. In addition, the visibility of some methods and classes has been changed. Select Strings previously hard coded are now declared as final class constants upfront resulting in a cleaner and more maintainable code. Furthermore, internal 'if' flow in 'onCommand' and 'onTabComplete' methods has been enhanced for better readability. Also, checks are added in certain places in the code to avoid NullPointerException.